+++ workspace/solutions/kotlin/lib/src/test/kotlin/s0/s4/s9/s6/s_0496_test.kt
@@ -1,0 +1,16 @@
+package s0.s4.s9.s6
+
+import utils.T
+import kotlin.test.Test
+import s0.s4.s9.s6.s_0496_00.Solution as s_0496_00
+
+class s_0496_test {
+    @Test
+    fun test() {
+        val t = T(
+            this::class,
+            s_0496_00::class,
+        )
+        t.run()
+    }
+}