+++ workspace/solutions/rust/src/s0/s4/s9/s6/mod.rs
@@ -1,0 +1,4 @@
+pub mod s_0496_00;
+pub mod s_0496_test;
+
+pub type TSolution = fn(Vec<i32>, Vec<i32>) -> Vec<i32>;